requirements.txt
Flask==0.9
Jinja2==2.6
Mako==0.7.3
MarkupSafe==0.15
SQLAlchemy==0.8.0
Werkzeug==0.8.3
alembic==0.4.2
nose==1.2.1
passlib==1.6.1
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2
